requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
protobuf==3.17.3
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.23
urllib3==1.26.6
Werkzeug==2.0.1
